* Merge from SysteMoC--devel--0.9 branch - Fixup some compiler warnings.
Patches applied:

 * falk@cs.fau.de--2008-systemoc-intern/SysteMoC--devel--0.9--patch-110
   
 * Fixup some compiler warnings.


git-archimport-id: falk@cs.fau.de--2008-systemoc-intern/SysteMoC--mainline--0.9.1--patch-40

+++ pkginclude/systemoc/smoc_multiplex_fifo.hpp
@@ -690,7 +690,7 @@ protected:
 
   /// @brief See smoc_port_in_base_if
   size_t inTokenId() const
-    { return -1; }
+    { return static_cast<size_t>(-1); }
 
   /// @brief See smoc_port_in_if
   AccessImpl *getReadPortAccess()
